@@ -12,14 +12,13 @@
     "@types/react-dom": "^17.0.11",
     "axios": "^0.25.0",
     "city-timezones": "^1.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-google-login": "^5.2.2",
     "react-scripts": "5.0.0",
     "react-slick": "^0.28.1",
     "react-spinners": "^0.11.0",
     "react-toastify": "^9.0.7",
-    "react-toastr": "^3.0.0",
     "rxjs": "^7.5.2",
     "slick-carousel": "^1.8.1",
     "typescript": "^4.5.4",
