@@ -2,10 +2,10 @@ import React, {useContext, useEffect, useState} from 'react';
 import './header.css';
 import GoogleLogin from 'react-google-login';
 import Modal from "./modal";
-import SavedRegions from '../../share/saved-regions/saved-regions.js';
-import RegionService from "../../services/region-service";
-import SearchContext from "../../contexts/search-context";
-import SearchTermContext from '../../contexts/search-term-context';
+import SavedRegions from '../../share/saved-regions/saved-regions';
+import RegionService from "../../services/region-service.js";
+import SearchContext from "../../contexts/search-context.js";
+import SearchTermContext from '../../contexts/search-term-context.js';
 import {toast} from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import ColorThemeContext from '../../contexts/color-theme-context.js';
