@@ -2,12 +2,12 @@ import React, {useState} from 'react';
 import './app.css';
 import Header from '../../home_page/header-component/header';
 import RegionArea from '../../home_page/region-area/region-area';
-import RegionService from '../../services/region-service';
-import SavedRegions from '../saved-regions/saved-regions.js';
-import RegionContext from '../../contexts/region-context';
-import SearchContext from '../../contexts/search-context';
+import RegionService from '../../services/region-service.js';
+import SavedRegions from '../saved-regions/saved-regions';
+import RegionContext from '../../contexts/region-context.js';
+import SearchContext from '../../contexts/search-context.js';
 import SearchComponent from '../../search_page/search-component/search-component';
-import SearchTermContext from '../../contexts/search-term-context';
+import SearchTermContext from '../../contexts/search-term-context.js';
 import { ToastContainer, toast, Zoom } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import ColorThemeContext from '../../contexts/color-theme-context.js';
